// app/lib/types.ts
export type RootStackParamList = {
	Home: undefined;
	Quiz: { category: string };
	Result: { score: number; totalQuestions: number; quizId: string };
	Profile: undefined;
	DailyTasks: undefined;
	Leaderboard: undefined;
	Settings: undefined;
	Login: undefined;
	Register: undefined;
	ForgotPassword: undefined;
	ResetPassword: undefined;
	OnBoard: undefined;
	CreateAccount: undefined;
	Bookmarks: undefined;
	WordGame: undefined;
};